refactor: Split test_dependencies.py into focused test modules
Break up the 970-line test_dependencies.py into 4 smaller, cohesive files:
- test_type_detection.py: type detection functions (is_annotated_type, etc.)
- test_extraction.py: signature extraction, validation, alias extractors
- test_injection.py: Maybe/Required annotations, custom DI extractors
- test_annotation_conversion.py: type conversion, unions, complex types

All 61 tests preserved. Reduces merge conflicts and improves discoverability.
